Fix massive root filesystem corruption under some build conditions.
Sounds scary, but it happens at build time, not runtime, so it cannot
brick devices. However, it was a frustrating exercise to uncover it,
as it is not entirely deterministic. This has been there for quite a
long time and I honestly cannot tell why it happened to me now.
Perhaps the exact location of the sparse holes on the filesystem makes
a difference. Or perhaps some versions of mkfs can handle it, others
not. All I know for sure, is that with `sparse` enabled I got massive
filesystem corruption (though sane enough to be somewhat readable) on
the rootfs. Without it, it's clean.

Changelog: Title

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>